training/requirements.txt
torch
torchvision
transformers
accelerate
safetensors
diffusers
pillow